src/ai_serever/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-dotenv==1.0.0
httpx==0.26.0

# LangChain
langchain==0.1.0
langchain-openai==0.0.5
langchain-core==0.1.23
langchain-community==0.0.20
langchain_naver
